Remove old cache and populate with new cache

+++ website/cache_data.py
@@ -19,7 +19,8 @@ def cache_featured_playlist():
 def save_cache(data, file_name):
     if not os.path.isdir(ROOT_CACHE):
         os.makedirs(ROOT_CACHE)
-    os.remove(ROOT_CACHE + '/' + file_name)
+    if os.path.isfile(ROOT_CACHE+'/'+file_name):
+        os.remove(ROOT_CACHE + '/' + file_name)
     with open(ROOT_CACHE + '/' + file_name, 'w') as fpfile:
         fpfile.write(data)
